Move conversions to separate modules
